backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
flask-marshmallow==1.3.0
marshmallow-sqlalchemy==1.4.2
marshmallow==4.2.0
flask-cors==6.0.2
PyJWT==2.10.1
python-dotenv==1.2.1
PyMySQL==1.1.2
flasgger==0.9.7.1